refactor: move trace_io to utils for cross-component sharing
- Relocate trace_io module from data/ to utils/
- Update imports in scheduler/strategies.py
- Update imports in benchmark/profiles.py
- Update imports in data/deserializers/trace_synthetic.py
- Update imports in tests/unit/scheduler/test_trace_replay.py
