+++ src/guidellm/scheduler/strategies.py
@@ -26,7 +26,6 @@ from typing import Annotated, ClassVar, Literal, TypeVar
 
 from pydantic import Field, NonNegativeFloat, NonNegativeInt, PositiveInt, PrivateAttr
 
-from guidellm.data.trace_io import load_relative_timestamps
 from guidellm.schemas import PydanticClassRegistryMixin, RequestInfo
 from guidellm.utils.mixins import InfoMixin
 
@@ -40,7 +39,6 @@ __all__ = [
     "SynchronousStrategy",
     "ThroughputStrategy",
     "TraceReplayStrategy",
-    "load_relative_timestamps",
 ]
