+++ src/guidellm/utils/trace_io.py
